Renamed Wifi Server dir
